requirements.txt
bcrypt==3.2.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Pillow==8.2.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
WTForms==2.3.3